semana-02/dia05/requirements.txt
### pip install -r requirements.txt
flask
firebase_admin
python-dotenv

flask-bootstrap ## integra bootstrap con flask
flask-WTF ## What the form

pyrebase4